mountpoint-select-dialog: don't show the last entry twice
Currently, the mountpoint selection dialog shows the last entry in the list twice.

This is caused by incorrect EOF handling:
EOF is hit while trying to do formatted input after the last complete line is read.
Formatted input doesn't get anything from the stream,
but there are no checks for that,
so the loop continues with values from the previous iteration.

Check if `m_mountpoint` is empty before proceeding with statfs() and friends.